model-service/requirements.txt
flask
tensorflow
pillow
numpy
gunicorn
flask-restx
flask-cors